Multi_Agent/Streamlit_app_stable/requirements.txt
crewai
crewai_tools
# LLM Connectors
langchain-groq
litellm>=1.75.9
litellm-proxy-extras>=0.4.16
apscheduler
fastapi
uvicorn
# UI and Enviroment
python-dotenv
streamlit